Completion of adding the ability to zoom into movie posters with gestures
