+++ .env
@@ -1,0 +1,1 @@
+TMDB_API_KEY='eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiI1MDM1NDRiZWI1ZTE1ZjY1YTk0YmRiM2JhOGU5ZTQwMSIsIm5iZiI6MTczMzA2NTM3Ni45MzI5OTk4LCJzdWIiOiI2NzRjN2FhMDY1NGVlNjE0NGFjMWNlMWYiLCJzY29wZXMiOlsiYXBpX3JlYWQiXSwidmVyc2lvbiI6MX0.ykLp7HHmr8PD_KhQIFZ2xn-ExuB4JI1oZxrXTlJRPo0'
