+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "my-movie-dairy",
-    "slug": "my-movie-dairy",
+    "name": "my_app",
+    "slug": "my_app",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
