+++ babel.config.js
@@ -1,10 +1,15 @@
-module.exports = {
-    presets: ['module:metro-react-native-babel-preset'],
-    plugins: [
-        ['module:react-native-dotenv', {
-            moduleName: '@env',
-            path: '.env',
-            allowUndefined: false,
-        }],
-    ],
-};  
+module.exports = function (api) {
+    api.cache(true);
+    return {
+        presets: ['babel-preset-expo'],
+        plugins: [
+            [
+                'module:react-native-dotenv',
+                {
+                    moduleName: '@env',
+                    path: '.env',
+                },
+            ],
+        ],
+    };
+};
