+++ components/Footer/index.js
@@ -1,14 +1,17 @@
 import React from 'react';
-import { View, Text, StyleSheet, Image } from 'react-native';
+import { View, Text, StyleSheet, Dimensions, Platform } from 'react-native';
 
 const Footer = () => {
+    const isWeb = Platform.OS === 'web';
+    const { width } = Dimensions.get('window');
+
     return (
         <View style={styles.footerContainer}>
             <View style={styles.textContainer}>
                 <Text style={styles.footerText}>Desenvolvido por Gabriel Carvalho</Text>
             </View>
 
-            <View style={styles.textContainer}>
+            <View style={[styles.textContainer, { width: isWeb ? 'auto' : width * 0.8 }]}>
                 <Text style={styles.footerText}>Powered by https://www.themoviedb.org/</Text>
             </View>
         </View>
@@ -17,20 +20,23 @@ const Footer = () => {
 
 const styles = StyleSheet.create({
     footerContainer: {
-        height: 40,
-        display: 'flex',
         flexDirection: 'row',
         justifyContent: 'space-between',
-        backgroundColor: 'black'
+        alignItems: 'center',
+        backgroundColor: 'black',
+        paddingHorizontal: 10,
+        paddingVertical: 5,
     },
     textContainer: {
-        marginHorizontal: 50,
-        marginVertical: 10,
+        flex: 1,
+        marginHorizontal: 5,
     },
     footerText: {
         color: '#fff',
-        fontStyle: 'italic'
-    }
+        fontStyle: 'italic',
+        textAlign: 'center', // Centraliza o texto para telas menores
+        flexWrap: 'wrap', // Permite quebra de linha
+    },
 });
 
 export default Footer;
