+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "my-movie-dairy",
+  "name": "my_app",
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
@@ -10,15 +10,13 @@
   },
   "dependencies": {
     "@expo/metro-runtime": "~4.0.0",
-    "@expo/vector-icons": "^14.0.4",
     "@react-native-community/datetimepicker": "^8.2.0",
-    "@react-native-picker/picker": "^2.9.0",
-    "@react-navigation/bottom-tabs": "^7.0.7",
-    "@react-navigation/drawer": "^7.0.7",
-    "@react-navigation/native": "^7.0.12",
-    "@react-navigation/native-stack": "^7.1.1",
-    "@react-navigation/stack": "^7.0.17",
-    "axios": "^1.7.8",
+    "@react-native-picker/picker": "^2.10.2",
+    "@react-navigation/bottom-tabs": "^7.1.3",
+    "@react-navigation/drawer": "^7.0.19",
+    "@react-navigation/native": "^7.0.13",
+    "@react-navigation/native-stack": "^7.1.14",
+    "@react-navigation/stack": "^7.0.18",
     "expo": "~52.0.11",
     "expo-status-bar": "~2.0.0",
     "react": "18.3.1",
@@ -26,17 +24,16 @@
     "react-native": "0.76.3",
     "react-native-dotenv": "^3.4.11",
     "react-native-gesture-handler": "^2.21.2",
+    "react-native-image-zoom-viewer": "^3.0.1",
     "react-native-paper": "^5.12.5",
-    "react-native-reanimated": "^3.16.2",
+    "react-native-reanimated": "^3.16.3",
     "react-native-safe-area-context": "^4.14.0",
     "react-native-screens": "^4.3.0",
     "react-native-vector-icons": "^10.2.0",
     "react-native-web": "~0.19.13"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "@react-native-community/cli": "^15.1.2",
-    "metro-react-native-babel-preset": "^0.77.0"
+    "@babel/core": "^7.20.0"
   },
   "private": true
 }
