+++ views/FavoritePage/index.js
