+++ views/SearchPage/index.js
